% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{shinyChess}
\alias{shinyChess}
\alias{app}
\title{A Shiny application to play and study chess}
\usage{
app(port = 1997, host = "127.0.0.1")
}
\arguments{
\item{port}{The port to use for shiny, default is 1997}
}
\description{
A Shiny application to play and study chess
}
